# chenzekai_kama/KamaChat/api/v1/group_info_controller.go
package v1

import (
	"github.com/gin-gonic/gin"
	"kama_chat_server/internal/dto/request"
	"kama_chat_server/internal/service/gorm"
	"kama_chat_server/pkg/constants"
	"kama_chat_server/pkg/zlog"
	"net/http"
)

// CreateGroup 创建群聊
// @Summary 创建群聊
// @Tags group
// @Accept json
// @Produce json
// @Param body body request.CreateGroupRequest true "创建群聊请求"
// @Success 200 {object} SwaggerResponse
// @Router /group/createGroup [post]
func CreateGroup(c *gin.Context) {
	var createGroupReq request.CreateGroupRequest
	if err := c.BindJSON(&createGroupReq); err != nil {
		zlog.Error(err.Error())
		c.JSON(http.StatusOK, gin.H{
			"code":    500,
			"message": constants.SYSTEM_ERROR,
		})
		return
	}
	message, ret := gorm.GroupInfoService.CreateGroup(createGroupReq)
	JsonBack(c, message, ret, nil)
}

// LoadMyGroup 获取我创建的群聊
// @Summary 获取我创建的群聊
// @Tags group
// @Accept json
// @Produce json
// @Param body body request.OwnlistRequest true "用户ID"
// @Success 200 {object} SwaggerResponse
// @Router /group/loadMyGroup [post]
func LoadMyGroup(c *gin.Context) {
	var loadMyGroupReq request.OwnlistRequest
	if err := c.BindJSON(&loadMyGroupReq); err != nil {
		zlog.Error(err.Error())
		c.JSON(http.StatusOK, gin.H{
			"code":    500,
			"message": constants.SYSTEM_ERROR,
		})
		return
	}
	message, groupList, ret := gorm.GroupInfoService.LoadMyGroup(loadMyGroupReq.OwnerId)
	JsonBack(c, message, ret, groupList)
}

// CheckGroupAddMode 检查群聊加群方式
// @Summary 检查群聊加群方式
// @Tags group
// @Accept json
// @Produce json
// @Param body body request.CheckGroupAddModeRequest true "群聊ID"
// @Success 200 {object} SwaggerResponse
// @Router /group/checkGroupAddMode [post]
func CheckGroupAddMode(c *gin.Context) {
	var req request.CheckGroupAddModeRequest
	if err := c.BindJSON(&req); err != nil {
		zlog.Error(err.Error())
		c.JSON(http.StatusOK, gin.H{
			"code":    500,
			"message": constants.SYSTEM_ERROR,
		})
		return
	}
	message, addMode, ret := gorm.GroupInfoService.CheckGroupAddMode(req.GroupId)
	JsonBack(c, message, ret, addMode)
}

// EnterGroupDirectly 直接进群
// @Summary 直接进群
// @Tags group
// @Accept json
// @Produce json
// @Param body body request.EnterGroupDirectlyRequest true "直接进群请求"
// @Success 200 {object} SwaggerResponse
// @Router /group/enterGroupDirectly [post]
func EnterGroupDirectly(c *gin.Context) {
	var req request.EnterGroupDirectlyRequest
	if err := c.BindJSON(&req); err != nil {
		zlog.Error(err.Error())
		c.JSON(http.StatusOK, gin.H{
			"code":    500,
			"message": constants.SYSTEM_ERROR,
		})
		return
	}
	message, ret := gorm.GroupInfoService.EnterGroupDirectly(req.OwnerId, req.ContactId)
	JsonBack(c, message, ret, nil)
}

// LeaveGroup 退群
// @Summary 退群
// @Tags group
// @Accept json
// @Produce json
// @Param body body request.LeaveGroupRequest true "退群请求"
// @Success 200 {object} SwaggerResponse
// @Router /group/leaveGroup [post]
func LeaveGroup(c *gin.Context) {
	var req request.LeaveGroupRequest
	if err := c.BindJSON(&req); err != nil {
		zlog.Error(err.Error())
		c.JSON(http.StatusOK, gin.H{
			"code":    500,
			"message": constants.SYSTEM_ERROR,
		})
		return
	}
	message, ret := gorm.GroupInfoService.LeaveGroup(req.UserId, req.GroupId)
	JsonBack(c, message, ret, nil)
}

// DismissGroup 解散群聊
// @Summary 解散群聊
// @Tags group
// @Accept json
// @Produce json
// @Param body body request.DismissGroupRequest true "解散群聊请求"
// @Success 200 {object} SwaggerResponse
// @Router /group/dismissGroup [post]
func DismissGroup(c *gin.Context) {
	var req request.DismissGroupRequest
	if err := c.BindJSON(&req); err != nil {
		zlog.Error(err.Error())
		c.JSON(http.StatusOK, gin.H{
			"code":    500,
			"message": constants.SYSTEM_ERROR,
		})
		return
	}
	message, ret := gorm.GroupInfoService.DismissGroup(req.OwnerId, req.GroupId)
	JsonBack(c, message, ret, nil)
}

// GetGroupInfo 获取群聊详情
// @Summary 获取群聊详情
// @Tags group
// @Accept json
// @Produce json
// @Param body body request.GetGroupInfoRequest true "群聊ID"
// @Success 200 {object} SwaggerResponse
// @Router /group/getGroupInfo [post]
func GetGroupInfo(c *gin.Context) {
	var req request.GetGroupInfoRequest
	if err := c.BindJSON(&req); err != nil {
		zlog.Error(err.Error())
		c.JSON(http.StatusOK, gin.H{
			"code":    500,
			"message": constants.SYSTEM_ERROR,
		})
		return
	}
	message, groupInfo, ret := gorm.GroupInfoService.GetGroupInfo(req.GroupId)
	JsonBack(c, message, ret, groupInfo)
}

// GetGroupInfoList 获取群聊列表 - 管理员
// @Summary 获取群聊列表
// @Tags group
// @Accept json
// @Produce json
// @Success 200 {object} SwaggerResponse
// @Router /group/getGroupInfoList [post]
func GetGroupInfoList(c *gin.Context) {
	message, groupList, ret := gorm.GroupInfoService.GetGroupInfoList()
	JsonBack(c, message, ret, groupList)
}

// DeleteGroups 删除列表中群聊 - 管理员
// @Summary 删除群聊
// @Tags group
// @Accept json
// @Produce json
// @Param body body request.DeleteGroupsRequest true "群聊ID列表"
// @Success 200 {object} SwaggerResponse
// @Router /group/deleteGroups [post]
func DeleteGroups(c *gin.Context) {
	var req request.DeleteGroupsRequest
	if err := c.BindJSON(&req); err != nil {
		zlog.Error(err.Error())
		c.JSON(http.StatusOK, gin.H{
			"code":    500,
			"message": constants.SYSTEM_ERROR,
		})
		return
	}
	message, ret := gorm.GroupInfoService.DeleteGroups(req.UuidList)
	JsonBack(c, message, ret, nil)
}

// SetGroupsStatus 设置群聊是否启用
// @Summary 设置群聊状态
// @Tags group
// @Accept json
// @Produce json
// @Param body body request.SetGroupsStatusRequest true "群聊状态请求"
// @Success 200 {object} SwaggerResponse
// @Router /group/setGroupsStatus [post]
func SetGroupsStatus(c *gin.Context) {
	var req request.SetGroupsStatusRequest
	if err := c.BindJSON(&req); err != nil {
		zlog.Error(err.Error())
		c.JSON(http.StatusOK, gin.H{
			"code":    500,
			"message": constants.SYSTEM_ERROR,
		})
		return
	}
	message, ret := gorm.GroupInfoService.SetGroupsStatus(req.UuidList, req.Status)
	JsonBack(c, message, ret, nil)
}

// UpdateGroupInfo 更新群聊消息
// @Summary 更新群聊信息
// @Tags group
// @Accept json
// @Produce json
// @Param body body request.UpdateGroupInfoRequest true "更新群聊信息请求"
// @Success 200 {object} SwaggerResponse
// @Router /group/updateGroupInfo [post]
func UpdateGroupInfo(c *gin.Context) {
	var req request.UpdateGroupInfoRequest
	if err := c.BindJSON(&req); err != nil {
		zlog.Error(err.Error())
		c.JSON(http.StatusOK, gin.H{
			"code":    500,
			"message": constants.SYSTEM_ERROR,
		})
		return
	}
	message, ret := gorm.GroupInfoService.UpdateGroupInfo(req)
	JsonBack(c, message, ret, nil)
}

// GetGroupMemberList 获取群聊成员列表
// @Summary 获取群聊成员列表
// @Tags group
// @Accept json
// @Produce json
// @Param body body request.GetGroupMemberListRequest true "群聊ID"
// @Success 200 {object} SwaggerResponse
// @Router /group/getGroupMemberList [post]
func GetGroupMemberList(c *gin.Context) {
	var req request.GetGroupMemberListRequest
	if err := c.BindJSON(&req); err != nil {
		zlog.Error(err.Error())
		c.JSON(http.StatusOK, gin.H{
			"code":    500,
			"message": constants.SYSTEM_ERROR,
		})
		return
	}
	message, groupMemberList, ret := gorm.GroupInfoService.GetGroupMemberList(req.GroupId)
	JsonBack(c, message, ret, groupMemberList)
}

// RemoveGroupMembers 移除群聊成员
// @Summary 移除群聊成员
// @Tags group
// @Accept json
// @Produce json
// @Param body body request.RemoveGroupMembersRequest true "移除群聊成员请求"
// @Success 200 {object} SwaggerResponse
// @Router /group/removeGroupMembers [post]
func RemoveGroupMembers(c *gin.Context) {
	var req request.RemoveGroupMembersRequest
	if err := c.BindJSON(&req); err != nil {
		zlog.Error(err.Error())
		c.JSON(http.StatusOK, gin.H{
			"code":    500,
			"message": constants.SYSTEM_ERROR,
		})
		return
	}
	message, ret := gorm.GroupInfoService.RemoveGroupMembers(req)
	JsonBack(c, message, ret, nil)
}