# chenzekai_kama/KamaChat/docs/swagger.yaml
basePath: /
definitions:
...
  request.CheckGroupAddModeRequest:
    properties:
      group_id:
        type: string
    type: object
  request.CreateGroupRequest:
    properties:
      add_mode:
        type: integer
      avatar:
        type: string
      name:
        type: string
      notice:
        type: string
      owner_id:
        type: string
    type: object
...
  request.DeleteGroupsRequest:
    properties:
      uuid_list:
        items:
          type: string
        type: array
    type: object
...
  request.DismissGroupRequest:
    properties:
      group_id:
        type: string
      owner_id:
        type: string
    type: object
  request.EnterGroupDirectlyRequest:
    properties:
      contact_id:
        type: string
      owner_id:
        type: string
    type: object
...
  request.GetCurContactListInChatRoomRequest:
    properties:
      contact_id:
        type: string
      owner_id:
        type: string
    type: object
  request.GetGroupInfoRequest:
    properties:
      group_id:
        type: string
    type: object
  request.GetGroupMemberListRequest:
    properties:
      group_id:
        type: string
    type: object
  request.GetGroupMessageListRequest:
    properties:
      group_id:
        type: string
    type: object
  request.GetMessageListRequest:
    properties:
      user_one_id:
        type: string
      user_two_id:
        type: string
    type: object
...
  request.LeaveGroupRequest:
    properties:
      group_id:
        type: string
      user_id:
        type: string
    type: object
...
  request.OwnlistRequest:
    properties:
      owner_id:
        type: string
    type: object
...
  request.RemoveGroupMembersRequest:
    properties:
      group_id:
        type: string
      owner_id:
        type: string
      uuid_list:
        items:
          type: string
        type: array
    type: object
...
  request.SetGroupsStatusRequest:
    properties:
      status:
        type: integer
      uuid_list:
        items:
          type: string
        type: array
    type: object
...
  request.UpdateGroupInfoRequest:
    properties:
      add_mode:
        type: integer
      avatar:
        type: string
      name:
        type: string
      notice:
        type: string
      owner_id:
        type: string
      uuid:
        type: string
    type: object
...
  v1.SwaggerResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: ok
        type: string
    type: object
info:
  contact: {}
  description: KamaChat server API documentation.
  title: KamaChat API
  version: "1.0"
paths:
  /chatroom/getCurContactListInChatRoom:
    post:
      consumes:
      - application/json
      parameters:
      - description: 获取聊天室联系人列表请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.GetCurContactListInChatRoomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.SwaggerResponse'
      summary: 获取当前聊天室联系人列表
      tags:
      - chatroom
...
  /group/checkGroupAddMode:
    post:
      consumes:
      - application/json
      parameters:
      - description: 群聊ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.CheckGroupAddModeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.SwaggerResponse'
      summary: 检查群聊加群方式
      tags:
      - group
  /group/createGroup:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建群聊请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.CreateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.SwaggerResponse'
      summary: 创建群聊
      tags:
      - group
  /group/deleteGroups:
    post:
      consumes:
      - application/json
      parameters:
      - description: 群聊ID列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.DeleteGroupsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.SwaggerResponse'
      summary: 删除群聊
      tags:
      - group
  /group/dismissGroup:
    post:
      consumes:
      - application/json
      parameters:
      - description: 解散群聊请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.DismissGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.SwaggerResponse'
      summary: 解散群聊
      tags:
      - group
  /group/enterGroupDirectly:
    post:
      consumes:
      - application/json
      parameters:
      - description: 直接进群请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.EnterGroupDirectlyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.SwaggerResponse'
      summary: 直接进群
      tags:
      - group
  /group/getGroupInfo:
    post:
      consumes:
      - application/json
      parameters:
      - description: 群聊ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.GetGroupInfoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.SwaggerResponse'
      summary: 获取群聊详情
      tags:
      - group
...
  /group/getGroupMemberList:
    post:
      consumes:
      - application/json
      parameters:
      - description: 群聊ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.GetGroupMemberListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.SwaggerResponse'
      summary: 获取群聊成员列表
      tags:
      - group
  /group/leaveGroup:
    post:
      consumes:
      - application/json
      parameters:
      - description: 退群请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.LeaveGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.SwaggerResponse'
      summary: 退群
      tags:
      - group
  /group/loadMyGroup:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.OwnlistRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.SwaggerResponse'
      summary: 获取我创建的群聊
      tags:
      - group
  /group/removeGroupMembers:
    post:
      consumes:
      - application/json
      parameters:
      - description: 移除群聊成员请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.RemoveGroupMembersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.SwaggerResponse'
      summary: 移除群聊成员
      tags:
      - group
  /group/setGroupsStatus:
    post:
      consumes:
      - application/json
      parameters:
      - description: 群聊状态请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.SetGroupsStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.SwaggerResponse'
      summary: 设置群聊状态
      tags:
      - group
  /group/updateGroupInfo:
    post:
      consumes:
      - application/json
      parameters:
      - description: 更新群聊信息请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.UpdateGroupInfoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.SwaggerResponse'
      summary: 更新群聊信息
      tags:
      - group
...
  /message/getGroupMessageList:
    post:
      consumes:
      - application/json
      parameters:
      - description: 获取群聊消息记录请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.GetGroupMessageListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.SwaggerResponse'
      summary: 获取群聊消息记录
      tags:
      - message
  /message/getMessageList:
    post:
      consumes:
      - application/json
      parameters:
      - description: 获取聊天记录请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.GetMessageListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.SwaggerResponse'
      summary: 获取聊天记录
      tags:
      - message
  /message/uploadAvatar:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 头像文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.SwaggerResponse'
      summary: 上传头像
      tags:
      - message
  /message/uploadFile:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.SwaggerResponse'
      summary: 上传文件
      tags:
      - message
...
schemes:
- https
- http
swagger: "2.0"